Add helper for unimplemented services; refactor files a bit.
